@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-# Copyright 2021 - 2023 Universität Tübingen, DKFZ, EMBL, and Universität zu Köln
+# Copyright 2021 - 2024 Universität Tübingen, DKFZ, EMBL, and Universität zu Köln
 # for the German Human Genome-Phenome Archive (GHGA)
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -24,7 +24,6 @@ import re
 import sys
 from datetime import date
 from pathlib import Path
-from typing import Optional, Union
 
 # root directory of the package:
 ROOT_DIR = Path(__file__).parent.parent.resolve()
@@ -34,42 +33,45 @@ GLOBAL_COPYRIGHT_FILE_PATH = ROOT_DIR / ".devcontainer" / "license_header.txt"
 
 # exclude files and dirs from license header check:
 EXCLUDE = [
-    ".devcontainer",
-    "eggs",
-    ".eggs",
-    "dist",
-    "build",
-    "develop-eggs",
-    "lib",
-    "lib62",
-    "parts",
-    "sdist",
-    "wheels",
-    "pip-wheel-metadata",
     ".coveragerc",
+    ".devcontainer",
+    ".editorconfig",
+    ".eggs",
     ".git",
     ".github",
     ".flake8",
     ".gitignore",
-    ".pylintrc",
-    ".ruff.toml",
-    ".ruff_cache",
-    "example_config.yaml",
-    "config_schema.json",
-    "LICENSE",  # is checked but not for the license header
-    ".pre-commit-config.yaml",
-    "docs",
-    ".vscode",
     ".mypy_cache",
     ".mypy.ini",
+    ".pylintrc",
     ".pytest_cache",
-    ".editorconfig",
+    ".ruff.toml",
+    ".ruff_cache",
     ".template/.static_files.txt",
     ".template/.static_files_ignore.txt",
     ".template/.mandatory_files.txt",
     ".template/.mandatory_files_ignore.txt",
     ".template/.deprecated_files.txt",
     ".template/.deprecated_files_ignore.txt",
+    ".tox",
+    ".venv",
+    ".vscode",
+    "eggs",
+    "build",
+    "config_schema.json",
+    "dist",
+    "docs",
+    "develop-eggs",
+    "example_config.yaml",
+    "htmlcov",
+    "lib",
+    "lib62",
+    "parts",
+    "pip-wheel-metadata",
+    "sdist",
+    "venv",
+    "wheels",
+    "LICENSE",  # is checked but not for the license header
 ]
 
 # exclude file by file ending from license header check:
@@ -82,6 +84,8 @@ EXCLUDE_ENDINGS = [
     "md",
     "pub",
     "pyc",
+    "pyd",
+    "typed",
     "sec",
     "toml",
     "txt",
@@ -111,8 +115,8 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License."""
 
-# A list of all chars that may be used to introduce a comment:
-COMMENT_CHARS = ["#"]
+# A list of strings that may be used to introduce a line comment:
+LINE_COMMENTS = ["#"]
 
 AUTHOR = """Universität Tübingen, DKFZ, EMBL, and Universität zu Köln
 for the German Human Genome-Phenome Archive (GHGA)"""
@@ -135,11 +139,11 @@ class GlobalCopyrightNotice:
     """
 
     def __init__(self):
-        self._text: Optional[str] = None
-        self._n_lines: Optional[int] = None
+        self._text: str | None = None
+        self._n_lines: int | None = None
 
     @property
-    def text(self) -> Optional[str]:
+    def text(self) -> str | None:
         return self._text
 
     @text.setter
@@ -162,7 +166,7 @@ class GlobalCopyrightNotice:
 class UnexpectedBinaryFileError(RuntimeError):
     """Thrown when trying to read a binary file."""
 
-    def __init__(self, file_path: Union[str, Path]):
+    def __init__(self, file_path: str | Path):
         message = f"The file could not be read because it is binary: {str(file_path)}"
         super().__init__(message)
 
@@ -176,13 +180,13 @@ def get_target_files(
     """Get target files that are not match the exclude conditions.
     Args:
         target_dir (pathlib.Path): The target dir to search.
-        exclude (List[str], optional):
+        exclude (list[str], optional):
             Overwrite default list of file/dir paths relative to
             the target dir that shall be excluded.
-        exclude_endings (List[str], optional):
+        exclude_endings (list[str], optional):
             Overwrite default list of file endings that shall
             be excluded.
-        exclude_pattern (List[str], optional):
+        exclude_pattern (list[str], optional):
             Overwrite default list of regex patterns match file path
             for exclusion.
     """
@@ -198,24 +202,25 @@ def get_target_files(
         file_
         for file_ in all_files
         if not (
-            any([file_.is_relative_to(excl) for excl in exclude_normalized])
-            or any([str(file_).endswith(ending) for ending in exclude_endings])
-            or any([re.match(pattern, str(file_)) for pattern in exclude_pattern])
+            any(file_.is_relative_to(excl) for excl in exclude_normalized)
+            or any(str(file_).endswith(ending) for ending in exclude_endings)
+            or any(re.match(pattern, str(file_)) for pattern in exclude_pattern)
         )
     ]
     return target_files
 
 
-def normalized_line(line: str, chars_to_trim: list[str] = COMMENT_CHARS) -> str:
-    norm_line = line.strip()
+def normalized_line(line: str, line_comments: list[str] = LINE_COMMENTS) -> str:
+    line = line.strip()
+    for line_comment in line_comments:
+        line_without_comment = line.removeprefix(line_comment)
+        if line_without_comment != line:
+            line = line_without_comment.lstrip()
+            break
+    return line
 
-    for char in chars_to_trim:
-        norm_line = norm_line.strip(char)
 
-    return norm_line.strip("\n").strip("\t").strip()
-
-
-def normalized_text(text: str, chars_to_trim: list[str] = COMMENT_CHARS) -> str:
+def normalized_text(text: str, line_comments: list[str] = LINE_COMMENTS) -> str:
     "Normalize a license header text."
     lines = text.split("\n")
 
@@ -227,7 +232,7 @@ def normalized_text(text: str, chars_to_trim: list[str] = COMMENT_CHARS) -> str:
         if stripped_line.startswith("#!"):
             continue
 
-        norm_line = normalized_line(stripped_line)
+        norm_line = normalized_line(stripped_line, line_comments=line_comments)
 
         # exclude empty lines:
         if norm_line == "":
@@ -245,22 +250,17 @@ def format_copyright_template(copyright_template: str, author: str) -> str:
     return normalized_text(copyright_template.replace("{author}", author))
 
 
-def is_commented_line(line: str, comment_chars: list[str] = COMMENT_CHARS) -> bool:
+def is_commented_line(line: str, line_comments: list[str] = LINE_COMMENTS) -> bool:
     """Checks whether a line is a comment."""
-    line_stripped = line.strip()
-    for comment_char in comment_chars:
-        if line_stripped.startswith(comment_char):
-            return True
-
-    return False
+    return line.lstrip().startswith(tuple(line_comments))
 
 
 def is_empty_line(line: str) -> bool:
     """Checks whether a line is empty."""
-    return line.strip("\n").strip("\t").strip() == ""
+    return not line.strip()
 
 
-def get_header(file_path: Path, comment_chars: list[str] = COMMENT_CHARS):
+def get_header(file_path: Path, line_comments: list[str] = LINE_COMMENTS):
     """Extracts the header from a file and normalizes it."""
     header_lines: list[str] = []
 
@@ -268,7 +268,7 @@ def get_header(file_path: Path, comment_chars: list[str] = COMMENT_CHARS):
         with open(file_path) as file:
             for line in file:
                 if is_commented_line(
-                    line, comment_chars=comment_chars
+                    line, line_comments=line_comments
                 ) or is_empty_line(line):
                     header_lines.append(line)
                 else:
@@ -278,7 +278,7 @@ def get_header(file_path: Path, comment_chars: list[str] = COMMENT_CHARS):
 
     # normalize the lines:
     header = "".join(header_lines)
-    return normalized_text(header, chars_to_trim=comment_chars)
+    return normalized_text(header, line_comments=line_comments)
 
 
 def validate_year_string(year_string: str, min_year: int = MIN_YEAR) -> bool:
@@ -291,8 +291,8 @@ def validate_year_string(year_string: str, min_year: int = MIN_YEAR) -> bool:
     if year_string.isnumeric():
         return int(year_string) == current_year
 
-    # Otherwise, a range (e.g. 2021 - 2023) is expected:
-    match = re.match("(\d+) - (\d+)", year_string)
+    # Otherwise, a range (e.g. 2021 - 2024) is expected:
+    match = re.match(r"(\d+) - (\d+)", year_string)
 
     if not match:
         return False
@@ -313,7 +313,7 @@ def check_copyright_notice(
     global_copyright: GlobalCopyrightNotice,
     copyright_template: str = COPYRIGHT_TEMPLATE,
     author: str = AUTHOR,
-    comment_chars: list[str] = COMMENT_CHARS,
+    line_comments: list[str] = LINE_COMMENTS,
     min_year: int = MIN_YEAR,
 ) -> bool:
     """Checks the specified copyright text against a template.
@@ -381,7 +381,7 @@ def check_file_headers(
     exclude: list[str] = EXCLUDE,
     exclude_endings: list[str] = EXCLUDE_ENDINGS,
     exclude_pattern: list[str] = EXCLUDE_PATTERN,
-    comment_chars: list[str] = COMMENT_CHARS,
+    line_comments: list[str] = LINE_COMMENTS,
     min_year: int = MIN_YEAR,
 ) -> tuple[list[Path], list[Path]]:
     """Check files for presence of a license header and verify that
@@ -402,13 +402,13 @@ def check_file_headers(
             The author that shall be included in the license header.
             It will replace any appearance of "{author}" in the license
             header. This defaults to an author info for GHGA.
-        exclude (List[str], optional):
+        exclude (list[str], optional):
             Overwrite default list of file/dir paths relative to
             the target dir that shall be excluded.
-        exclude_endings (List[str], optional):
+        exclude_endings (list[str], optional):
             Overwrite default list of file endings that shall
             be excluded.
-        exclude_pattern (List[str], optional):
+        exclude_pattern (list[str], optional):
             Overwrite default list of regex patterns match file path
             for exclusion.
     """
@@ -425,13 +425,13 @@ def check_file_headers(
 
     for target_file in target_files:
         try:
-            header = get_header(target_file, comment_chars=comment_chars)
+            header = get_header(target_file, line_comments=line_comments)
             if check_copyright_notice(
                 copyright=header,
                 global_copyright=global_copyright,
                 copyright_template=copyright_template,
                 author=author,
-                comment_chars=comment_chars,
+                line_comments=line_comments,
                 min_year=min_year,
             ):
                 passed_files.append(target_file)
@@ -449,7 +449,7 @@ def check_license_file(
     global_copyright: GlobalCopyrightNotice,
     copyright_template: str = COPYRIGHT_TEMPLATE,
     author: str = AUTHOR,
-    comment_chars: list[str] = COMMENT_CHARS,
+    line_comments: list[str] = LINE_COMMENTS,
     min_year: int = MIN_YEAR,
 ) -> bool:
     """Currently only checks if the copyright notice in the
@@ -491,7 +491,7 @@ def check_license_file(
         global_copyright=global_copyright,
         copyright_template=copyright_template,
         author=author,
-        comment_chars=comment_chars,
+        line_comments=line_comments,
         min_year=min_year,
     )
 
