@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-# Copyright 2021 - 2023 Universität Tübingen, DKFZ, EMBL, and Universität zu Köln
+# Copyright 2021 - 2024 Universität Tübingen, DKFZ, EMBL, and Universität zu Köln
 # for the German Human Genome-Phenome Archive (GHGA)
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -20,11 +20,11 @@
 import json
 import subprocess  # nosec
 import sys
+import tomllib
 from pathlib import Path
 from string import Template
 
 import jsonschema2md
-import tomli
 from pydantic import BaseModel, Field
 from stringcase import spinalcase, titlecase
 
@@ -97,7 +97,7 @@ def read_toml_package_header() -> PackageHeader:
     """Read basic information about the package from the pyproject.toml"""
 
     with open(PYPROJECT_TOML_PATH, "rb") as pyproject_toml:
-        pyproject = tomli.load(pyproject_toml)
+        pyproject = tomllib.load(pyproject_toml)
         pyproject_project = pyproject["project"]
         return PackageHeader(
             shortname=pyproject_project["name"],
