@@ -17,7 +17,7 @@
 
 import logging
 from enum import Enum
-from typing import Annotated, Union
+from typing import Annotated
 
 from fastapi import FastAPI, HTTPException, Response, Security, status
 from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer
@@ -35,7 +35,7 @@ configure_app(app, config=CONFIG)
 
 oidc_provider = OidcProvider(CONFIG)
 
-tags: list[Union[str, Enum]] = ["TestOP"]
+tags: list[str | Enum] = ["TestOP"]
 
 
 @app.get(
