@@ -16,8 +16,6 @@
 
 """Models used by the API"""
 
-from typing import Optional, Union
-
 from pydantic import (
     AnyHttpUrl,
     BaseModel,
@@ -36,12 +34,12 @@ __all__ = ["LoginInfo", "UserInfo", "OidcConfiguration"]
 class LoginInfo(BaseModel):
     """Data that is used to login as a user."""
 
-    sub: Optional[str] = Field(default=None, description="subject identifier")
-    email: Optional[EmailStr] = Field(
+    sub: str | None = Field(default=None, description="subject identifier")
+    email: EmailStr | None = Field(
         default=None, description="e-mail address of the user"
     )
     name: str = Field(..., description="the full name of the user")
-    valid_seconds: Optional[Union[PositiveInt, PositiveFloat]] = Field(
+    valid_seconds: PositiveInt | PositiveFloat | None = Field(
         default=None, description="seconds until the login expires"
     )
 
